+++ init/src/utils/containerPath.ts
@@ -1,0 +1,6 @@
+import os from "os";
+import path from "path";
+
+export const containerPath = path.join(os.homedir(), "Desktop/newFolder/folder/");
+
+console.log(containerPath);
